+++ spec/rubocop/cop/style/access_modifier_indentation_spec.rb
@@ -2,11 +2,11 @@
 
 require 'spec_helper'
 
-describe Rubocop::Cop::Style::AccessControl, :config do
+describe Rubocop::Cop::Style::AccessModifierIndentation, :config do
   subject(:cop) { described_class.new(config) }
 
   context 'when IndentDepth is set to method' do
-    let(:cop_config) { { 'IndentDepth' => 'method' } }
+    let(:cop_config) { { 'EnforcedStyle' => 'indent' } }
 
     it 'registers an offence for misaligned private' do
       inspect_source(cop,
@@ -18,7 +18,7 @@ describe Rubocop::Cop::Style::AccessControl, :config do
                       'end'])
       expect(cop.offences.size).to eq(1)
       expect(cop.messages)
-        .to eq(['Indent private as deep as method definitions.'])
+        .to eq(['Indent access modifiers like private.'])
     end
 
     it 'registers an offence for misaligned private in module' do
@@ -31,7 +31,7 @@ describe Rubocop::Cop::Style::AccessControl, :config do
                       'end'])
       expect(cop.offences.size).to eq(1)
       expect(cop.messages)
-        .to eq(['Indent private as deep as method definitions.'])
+        .to eq(['Indent access modifiers like private.'])
     end
 
     it 'registers an offence for misaligned private in singleton class' do
@@ -44,7 +44,7 @@ describe Rubocop::Cop::Style::AccessControl, :config do
                       'end'])
       expect(cop.offences.size).to eq(1)
       expect(cop.messages)
-        .to eq(['Indent private as deep as method definitions.'])
+        .to eq(['Indent access modifiers like private.'])
     end
 
     it 'registers an offence for misaligned private in class ' +
@@ -58,7 +58,7 @@ describe Rubocop::Cop::Style::AccessControl, :config do
                       'end'])
       expect(cop.offences.size).to eq(1)
       expect(cop.messages)
-        .to eq(['Indent private as deep as method definitions.'])
+        .to eq(['Indent access modifiers like private.'])
     end
 
     it 'registers an offence for misaligned private in module ' +
@@ -72,7 +72,7 @@ describe Rubocop::Cop::Style::AccessControl, :config do
                       'end'])
       expect(cop.offences.size).to eq(1)
       expect(cop.messages)
-        .to eq(['Indent private as deep as method definitions.'])
+        .to eq(['Indent access modifiers like private.'])
     end
 
     it 'registers an offence for misaligned protected' do
@@ -85,7 +85,7 @@ describe Rubocop::Cop::Style::AccessControl, :config do
                       'end'])
       expect(cop.offences.size).to eq(1)
       expect(cop.messages)
-        .to eq(['Indent protected as deep as method definitions.'])
+        .to eq(['Indent access modifiers like protected.'])
     end
 
     it 'accepts properly indented private' do
@@ -127,48 +127,13 @@ describe Rubocop::Cop::Style::AccessControl, :config do
                       'end'])
       expect(cop.offences.size).to eq(1)
       expect(cop.messages)
-        .to eq(['Indent private as deep as method definitions.'])
-    end
-
-    it 'requires blank line before private/protected' do
-      inspect_source(cop,
-                     ['class Test',
-                      '  protected',
-                      '',
-                      '  def test; end',
-                      'end'])
-      expect(cop.offences.size).to eq(1)
-      expect(cop.messages)
-        .to eq(['Keep a blank line before and after protected.'])
-    end
-
-    it 'requires blank line after private/protected' do
-      inspect_source(cop,
-                     ['class Test',
-                      '',
-                      '  protected',
-                      '  def test; end',
-                      'end'])
-      expect(cop.offences.size).to eq(1)
-      expect(cop.messages)
-        .to eq(['Keep a blank line before and after protected.'])
-    end
-
-    it 'recognizes blank lines with DOS style line endings' do
-      inspect_source(cop,
-                     ["class Test\r",
-                      "\r",
-                      "  protected\r",
-                      "\r",
-                      "  def test; end\r",
-                      "end\r"])
-      expect(cop.offences.size).to eq(0)
+        .to eq(['Indent access modifiers like private.'])
     end
   end
 
   context 'when IndentDepth is set to class' do
-    let(:cop_config) { { 'IndentDepth' => 'class' } }
-    let(:indent_msg) { 'Indent private as deep as class definitions.' }
+    let(:cop_config) { { 'EnforcedStyle' => 'outdent' } }
+    let(:indent_msg) { 'Outdent access modifiers like private.' }
     let(:blank_msg) { 'Keep a blank line before and after private.' }
 
     it 'registers offence for private indented to method depth in a class' do
@@ -273,39 +238,6 @@ describe Rubocop::Cop::Style::AccessControl, :config do
                       'end'])
       expect(cop.offences.size).to eq(1)
       expect(cop.messages).to eq([indent_msg])
-    end
-
-    it 'requires blank line before private/protected' do
-      inspect_source(cop,
-                     ['class Test',
-                      'private',
-                      '',
-                      '  def test; end',
-                      'end'])
-      expect(cop.offences.size).to eq(1)
-      expect(cop.messages).to eq([blank_msg])
-    end
-
-    it 'requires blank line after private/protected' do
-      inspect_source(cop,
-                     ['class Test',
-                      '',
-                      'private',
-                      '  def test; end',
-                      'end'])
-      expect(cop.offences.size).to eq(1)
-      expect(cop.messages).to eq([blank_msg])
-    end
-
-    it 'recognizes blank lines with DOS style line endings' do
-      inspect_source(cop,
-                     ["class Test\r",
-                      "\r",
-                      "protected\r",
-                      "\r",
-                      "  def test; end\r",
-                      "end\r"])
-      expect(cop.offences.size).to eq(0)
     end
   end
 end