+++ CHANGELOG.md
@@ -27,7 +27,8 @@
 * [#557](https://github.com/bbatsov/rubocop/pull/557): Configuration files for excluded files are no longer loaded. ([@jonas054][])
 * [#571](https://github.com/bbatsov/rubocop/pull/571): The default rake task now runs RuboCop over its self! ([@nevir][])
 * Encoding errors are reported as fatal offences rather than printed with red text.
-* `AccessControl` cop is now configurable with the `IndentDepth` option
+* `AccessControl` cop is now configurable with the `EnforcedStyle` option.
+* Split `AccessControl` cop to `AccessModifierIndentation` and `EmptyLinesAroundAccessModifier`. ([@bbatsov][])
 
 ### Bugs fixed